refactored specs to put BooleanFromString based api parameters on a file not used by mobile team
