+++ api_cgn_operator_search.yaml
@@ -173,21 +173,6 @@ definitions:
     $ref: "https://raw.githubusercontent.com/pagopa/io-functions-cgn-operator-search/v2.6.2/openapi/index.yaml#/definitions/PublishedProductCategoriesResult"
   ProductCategoryWithNewDiscountsCount:
     $ref: "https://raw.githubusercontent.com/pagopa/io-functions-cgn-operator-search/v2.6.2/openapi/index.yaml#/definitions/ProductCategoryWithNewDiscountsCount"
-  GetPublishedCategoriesParameters:
-    type: object
-    title: GetPublishedCategoriesParameters
-    description: Describes the GetMessages api query parameters
-    properties:
-      count_new_discounts:
-        $ref: "#/definitions/BooleanFromString"
-  BooleanFromString:
-    type: string
-    description: Represents a boolean parsed from a string
-    format: BooleanFromString
-    # TODO: Import from ts-commons when it will be updated during the
-    # update of backend to fp-ts 2.x
-    x-import: '../../src/types/booleans'
-    example: true
 securityDefinitions:
   Bearer:
     type: apiKey
