+++ src/controllers/cgnOperatorSearchController.ts
@@ -22,7 +22,7 @@ import { Card } from "generated/cgn/Card";
 import { Merchant } from "../../generated/cgn-operator-search/Merchant";
 import { OfflineMerchants } from "../../generated/cgn-operator-search/OfflineMerchants";
 import { OnlineMerchants } from "../../generated/cgn-operator-search/OnlineMerchants";
-import { GetPublishedCategoriesParameters } from "../../generated/cgn-operator-search/GetPublishedCategoriesParameters";
+import { GetPublishedCategoriesParameters } from "../../generated/parameters/GetPublishedCategoriesParameters";
 import { PublishedProductCategoriesResult } from "../../generated/cgn-operator-search/PublishedProductCategoriesResult";
 import { OfflineMerchantSearchRequest } from "../../generated/io-cgn-operator-search-api/OfflineMerchantSearchRequest";
 import { OnlineMerchantSearchRequest } from "../../generated/io-cgn-operator-search-api/OnlineMerchantSearchRequest";
