+++ src/controllers/messagesController.ts
@@ -26,7 +26,7 @@ import { withUserFromRequest } from "../types/user";
 import { MessageStatusChange } from "../../generated/io-api/MessageStatusChange";
 import { MessageStatusAttributes } from "../../generated/io-api/MessageStatusAttributes";
 import { PaginatedPublicMessagesCollection } from "../../generated/backend/PaginatedPublicMessagesCollection";
-import { GetMessagesParameters } from "../../generated/backend/GetMessagesParameters";
+import { GetMessagesParameters } from "../../generated/parameters/GetMessagesParameters";
 import {
   withValidatedOrValidationError,
   IResponseSuccessOctet
